refactor: layout composition components
