--- src/components/Layout/compositions/index.ts
@@ -1,2 +1,0 @@
-export { Menu } from './Menu'
-export { Topbar } from './Topbar'
